allbarefoot_bot/requirements.txt
gino
wheel
asyncio
redis
aioredis
asyncpg
aiogram==2.25
xlsxwriter
python-docx
yahoo-fin
gspread-asyncio